+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 - `REPL`支持自动添加`print`
 - 不再支持任意对象 `concat`，如不能拼接则会返回 `""`
 - 去除 `metatable` 属性，直接将其置于 `luaTable` 对象中
+- 优化 `luaTable` 默认 `print` 格式
+- `PCall` 仅在 `REPL` 输出错误
 
 ## 0.1.2
 - `Table`索引从`0`开始